requirements.txt
absl-py==0.11.0
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
Brlapi==0.6.6
cachetools==4.2.1
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cx-Oracle==8.1.0
cycler==0.10.0
dataclasses==0.8
defer==1.0.6
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
Django==3.1.5
djangorestframework==3.12.2
facenet==1.0.5
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.9.2
idna==2.6
importlib-metadata==3.4.0
imutils==0.5.3
isort==5.7.0
joblib==1.0.0
Keras==2.2.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.3
MarkupSafe==1.0
matplotlib==3.3.4
mccabe==0.6.1
mediapipe==0.8.2
netifaces==0.10.4
numpy==1.19.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.0.1.23
opt-einsum==3.3.0
packaging==20.4
pexpect==4.2.1
Pillow==8.1.0
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7
scikit-build==0.11.1
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sqlparse==0.4.1
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.4.0
zope.interface==4.3.2
